[REFACTOR] Move new signing key types to separate module
This helps to clarify things, and keeps all the old code contained by itself.
