+++ components/common/src/command/package/install.rs
@@ -32,9 +32,9 @@ use crate::{api_client::{self,
                  UIWriter}};
 use habitat_core::{self,
                    crypto::{artifact,
-                            keys::{sig_key_pair::PublicOriginSigningKey,
-                                   KeyCache,
-                                   NamedRevision}},
+                            keys::{KeyCache,
+                                   NamedRevision,
+                                   PublicOriginSigningKey}},
                    fs::{cache_key_path,
                         pkg_install_path,
                         AtomicWriter,
