+++ components/core/src/crypto/keys/sig_key_pair.rs
@@ -8,169 +8,23 @@ use super::{super::{hash,
             parse_name_with_rev,
             write_keypair_files,
             HabitatKey,
-            Key,
             KeyPair,
             KeyRevision,
             KeyType,
             PairType,
             TmpKeyfile};
-use crate::{crypto::keys::NamedRevision,
-            error::{Error,
-                    Result}};
+use crate::error::{Error,
+                   Result};
 use sodiumoxide::{crypto::{sign,
                            sign::ed25519::{PublicKey as SigPublicKey,
                                            SecretKey as SigSecretKey}},
                   randombytes::randombytes};
 use std::{convert::TryFrom,
           fs,
-          io::Read,
           path::{Path,
                  PathBuf}};
 
-from_slice_impl_for_sodiumoxide_key!(SigPublicKey);
-from_slice_impl_for_sodiumoxide_key!(SigSecretKey);
-
 pub type SigKeyPair = KeyPair<SigPublicKey, SigSecretKey>;
-
-/// Given the name of an origin, generate a new signing key pair.
-///
-/// The resulting keys will need to be saved to a cache in order to
-/// persist.
-pub fn generate_signing_key_pair(origin_name: &str)
-                                 -> (PublicOriginSigningKey, SecretOriginSigningKey) {
-    let revision = KeyRevision::new();
-    let (pk, sk) = sign::gen_keypair();
-
-    let public = PublicOriginSigningKey::from_raw(origin_name.to_string(), revision.clone(), pk);
-    let secret = SecretOriginSigningKey::from_raw(origin_name.to_string(), revision.clone(), sk);
-    (public, secret)
-}
-
-////////////////////////////////////////////////////////////////////////
-
-pub struct PublicOriginSigningKey {
-    named_revision: NamedRevision,
-    key:            SigPublicKey,
-    path:           PathBuf,
-}
-
-impl Key for PublicOriginSigningKey {
-    type Crypto = SigPublicKey;
-
-    // PUBLIC_KEY_SUFFIX;
-    const EXTENSION: &'static str = "pub";
-    const PERMISSIONS: crate::fs::Permissions = crate::fs::DEFAULT_PUBLIC_KEY_PERMISSIONS;
-    const VERSION_STRING: &'static str = PUBLIC_SIG_KEY_VERSION;
-
-    fn key(&self) -> &SigPublicKey { &self.key }
-
-    fn named_revision(&self) -> &NamedRevision { &self.named_revision }
-}
-
-impl PublicOriginSigningKey {
-    pub(crate) fn from_raw(name: String, revision: KeyRevision, key: SigPublicKey) -> Self {
-        let named_revision = NamedRevision::new(name, revision);
-        let path = named_revision.filename::<Self>();
-        Self { named_revision,
-               key,
-               path }
-    }
-
-    // Simple helper to deal with the indirection to the inner
-    // KeyPair struct. Not ultimately sure if this should be kept.
-    pub fn name_with_rev(&self) -> String { self.named_revision.to_string() }
-
-    /// Accept a signature and the bytes for the signed content to be
-    /// verified. Returns the named revision of the key as well as the
-    /// computed hash.
-    // TODO (CM): Result should be NamedRevision, String
-    pub fn verify(&self, signature: &[u8], content: &mut dyn Read) -> Result<(String, String)> {
-        // TODO (CM): we should always have a public key here, by definition.
-        let expected_hash = match sign::verify(signature, &self.key) {
-            Ok(signed_data) => {
-                String::from_utf8(signed_data).map_err(|_| {
-                                                  Error::CryptoError("Error parsing artifact \
-                                                                      signature"
-                                                                                .to_string())
-                                              })?
-            }
-            Err(_) => return Err(Error::CryptoError("Verification failed".to_string())),
-        };
-
-        let computed_hash = hash::hash_reader(content)?;
-        if computed_hash == expected_hash {
-            Ok((self.name_with_rev(), expected_hash))
-        } else {
-            let msg = format!("Habitat artifact is invalid, hashes don't match (expected: {}, \
-                               computed: {})",
-                              expected_hash, computed_hash);
-            Err(Error::CryptoError(msg))
-        }
-    }
-}
-
-from_str_impl_for_key!(PublicOriginSigningKey);
-
-try_from_path_buf_impl_for_key!(PublicOriginSigningKey);
-
-as_ref_path_impl_for_key!(PublicOriginSigningKey);
-
-////////////////////////////////////////////////////////////////////////
-
-pub struct SecretOriginSigningKey {
-    named_revision: NamedRevision,
-    key:            SigSecretKey,
-    path:           PathBuf,
-}
-
-impl Key for SecretOriginSigningKey {
-    type Crypto = SigSecretKey;
-
-    // SECRET_SIG_KEY_SUFFIX;
-    const EXTENSION: &'static str = "sig.key";
-    const PERMISSIONS: crate::fs::Permissions = crate::fs::DEFAULT_SECRET_KEY_PERMISSIONS;
-    const VERSION_STRING: &'static str = SECRET_SIG_KEY_VERSION;
-
-    fn key(&self) -> &SigSecretKey { &self.key }
-
-    fn named_revision(&self) -> &NamedRevision { &self.named_revision }
-}
-
-type Signature = Vec<u8>;
-
-impl SecretOriginSigningKey {
-    pub(crate) fn from_raw(name: String, revision: KeyRevision, key: SigSecretKey) -> Self {
-        let named_revision = NamedRevision::new(name, revision);
-        let path = named_revision.filename::<Self>();
-        Self { named_revision,
-               key,
-               path }
-    }
-
-    // Simple helper to deal with the indirection to the inner
-    // KeyPair struct. Not ultimately sure if this should be kept.
-    pub fn name_with_rev(&self) -> String { self.named_revision.to_string() }
-
-    /// Takes the contents of the given file and returns a signature
-    /// based on this key.
-    pub fn sign<P>(&self, path: P) -> Result<Signature>
-        where P: AsRef<Path>
-    {
-        let hash = hash::hash_file(&path)?;
-        debug!("File hash for {} = {}", path.as_ref().display(), &hash);
-        Ok(sign::sign(&hash.as_bytes(), &self.key))
-    }
-}
-
-from_str_impl_for_key!(SecretOriginSigningKey);
-
-try_from_path_buf_impl_for_key!(SecretOriginSigningKey);
-
-as_ref_path_impl_for_key!(SecretOriginSigningKey);
-
-////////////////////////////////////////////////////////////////////////
-// OLD STUFF BELOW!
-////////////////////////////////////////////////////////////////////////
 
 impl SigKeyPair {
     //    #[cfg(test)]
@@ -465,6 +319,7 @@ mod test {
                 KeyRevision,
                 SigKeyPair,
                 *};
+    use crate::crypto::keys::generate_signing_key_pair;
     use std::{fs::{self,
                    File},
               io::Read};
